+++ libs/src/python/salmon_fs/salmon_fs/salmonfs/drive/utils/aes_file_importer.py
@@ -362,9 +362,11 @@ class AesFileImporter:
         """!
         Close this importer and associated resources
         """
-        self.__executor.shutdown()
-        self.__shm_cancel.close()
-        self.__shm_cancel.unlink()
+        if self.__executor:
+            self.__executor.shutdown(False)
+        if self.__shm_cancel:
+            self.__shm_cancel.close()
+            self.__shm_cancel.unlink()
 
     class FileImportOptions:
         """!
